First Commit
The Smart Attendance System using Face Recognition is an automated solution designed to streamline attendance tracking in educational or professional settings. It utilizes Python and OpenCV for face detection and recognition, leveraging the Haar Cascade Classifier for identifying faces and the Local Binary Patterns Histogram (LBPH) algorithm for efficient recognition. The system captures real-time images, processes them to extract facial features, and compares them with a stored database for identification. This ensures accurate attendance marking, reduces manual effort, and enhances security. The system is user-friendly, scalable, and suitable for real-world applications in schools, offices, and institutions.